Address vehicles block
