+++ src/OpenLoco/src/GameCommands/Vehicles/CloneVehicle.cpp
@@ -1,6 +1,11 @@
+#include "CloneVehicle.h"
+#include "CreateVehicle.h"
 #include "Entities/EntityManager.h"
 #include "GameCommands/GameCommands.h"
 #include "Ui/WindowManager.h"
+#include "VehicleChangeRunningMode.h"
+#include "VehicleOrderInsert.h"
+#include "VehicleRefit.h"
 #include "Vehicles/OrderManager.h"
 #include "Vehicles/Orders.h"
 #include "Vehicles/Vehicle.h"
