+++ src/OpenLoco/src/GameCommands/Vehicles/CreateVehicle.cpp
@@ -1,3 +1,4 @@
+#include "CreateVehicle.h"
 #include "Audio/Audio.h"
 #include "Config.h"
 #include "Date.h"
