+++ src/OpenLoco/src/GameCommands/Vehicles/RenameVehicle.cpp
@@ -1,3 +1,4 @@
+#include "RenameVehicle.h"
 #include "Economy/Expenditures.h"
 #include "Entities/EntityManager.h"
 #include "GameCommands/GameCommands.h"
