+++ src/OpenLoco/src/GameCommands/Vehicles/VehicleChangeRunningMode.cpp
@@ -1,9 +1,11 @@
+#include "VehicleChangeRunningMode.h"
 #include "Economy/Expenditures.h"
 #include "Entities/EntityManager.h"
 #include "GameCommands/GameCommands.h"
 #include "Speed.hpp"
 #include "Types.hpp"
 #include "Ui/WindowManager.h"
+#include "VehicleSell.h"
 #include "Vehicles/Vehicle.h"
 #include "World/CompanyManager.h"
 #include "World/StationManager.h"
