+++ src/OpenLoco/src/GameCommands/Vehicles/VehicleOrderDelete.cpp
@@ -1,3 +1,4 @@
+#include "VehicleOrderDelete.h"
 #include "Entities/EntityManager.h"
 #include "GameCommands/GameCommands.h"
 #include "Ui/WindowManager.h"
