+++ src/OpenLoco/src/GameCommands/Vehicles/VehicleOrderDown.cpp
@@ -1,3 +1,4 @@
+#include "VehicleOrderDown.h"
 #include "Entities/EntityManager.h"
 #include "GameCommands/GameCommands.h"
 #include "Ui/WindowManager.h"
