+++ src/OpenLoco/src/GameCommands/Vehicles/VehicleOrderReverse.cpp
@@ -1,3 +1,4 @@
+#include "VehicleOrderReverse.h"
 #include "Entities/EntityManager.h"
 #include "GameCommands/GameCommands.h"
 #include "Ui/WindowManager.h"
