+++ src/OpenLoco/src/GameCommands/Vehicles/VehicleOrderSkip.cpp
@@ -1,3 +1,4 @@
+#include "VehicleOrderSkip.h"
 #include "Entities/EntityManager.h"
 #include "GameCommands/GameCommands.h"
 #include "Ui/WindowManager.h"
