+++ src/OpenLoco/src/GameCommands/Vehicles/VehicleOrderUp.cpp
@@ -1,3 +1,4 @@
+#include "VehicleOrderUp.h"
 #include "Entities/EntityManager.h"
 #include "GameCommands/GameCommands.h"
 #include "Ui/WindowManager.h"
