+++ src/OpenLoco/src/GameCommands/Vehicles/VehiclePassSignal.cpp
@@ -1,3 +1,4 @@
+#include "VehiclePassSignal.h"
 #include "Economy/Expenditures.h"
 #include "Entities/EntityManager.h"
 #include "GameCommands/GameCommands.h"
