+++ src/OpenLoco/src/GameCommands/Vehicles/VehiclePickup.cpp
@@ -1,3 +1,4 @@
+#include "VehiclePickup.h"
 #include "Audio/Audio.h"
 #include "Economy/Expenditures.h"
 #include "Entities/EntityManager.h"
