+++ src/OpenLoco/src/GameCommands/Vehicles/VehiclePickupWater.h
@@ -4,5 +4,25 @@
 
 namespace OpenLoco::GameCommands
 {
+    struct VehiclePickupWaterArgs
+    {
+        static constexpr auto command = GameCommand::vehiclePickupWater;
+
+        VehiclePickupWaterArgs() = default;
+        explicit VehiclePickupWaterArgs(const registers& regs)
+            : head(static_cast<EntityId>(regs.di))
+        {
+        }
+
+        EntityId head;
+
+        explicit operator registers() const
+        {
+            registers regs;
+            regs.di = enumValue(head);
+            return regs;
+        }
+    };
+
     void vehiclePickupWater(Interop::registers& regs);
 }
