+++ src/OpenLoco/src/GameCommands/Vehicles/VehicleRefit.cpp
@@ -1,3 +1,4 @@
+#include "VehicleRefit.h"
 #include "Economy/Expenditures.h"
 #include "Entities/EntityManager.h"
 #include "GameCommands/GameCommands.h"
