+++ src/OpenLoco/src/GameCommands/Vehicles/VehicleReverse.cpp
@@ -1,3 +1,4 @@
+#include "VehicleReverse.h"
 #include "Economy/Expenditures.h"
 #include "Entities/EntityManager.h"
 #include "GameCommands/GameCommands.h"
