+++ src/OpenLoco/src/Windows/Vehicle.cpp
@@ -3,6 +3,22 @@
 #include "Drawing/SoftwareDrawingEngine.h"
 #include "Entities/EntityManager.h"
 #include "GameCommands/GameCommands.h"
+#include "GameCommands/Vehicles/RenameVehicle.h"
+#include "GameCommands/Vehicles/VehicleChangeRunningMode.h"
+#include "GameCommands/Vehicles/VehicleOrderDelete.h"
+#include "GameCommands/Vehicles/VehicleOrderDown.h"
+#include "GameCommands/Vehicles/VehicleOrderInsert.h"
+#include "GameCommands/Vehicles/VehicleOrderReverse.h"
+#include "GameCommands/Vehicles/VehicleOrderSkip.h"
+#include "GameCommands/Vehicles/VehicleOrderUp.h"
+#include "GameCommands/Vehicles/VehiclePassSignal.h"
+#include "GameCommands/Vehicles/VehiclePickup.h"
+#include "GameCommands/Vehicles/VehiclePickupAir.h"
+#include "GameCommands/Vehicles/VehiclePickupWater.h"
+#include "GameCommands/Vehicles/VehicleRefit.h"
+#include "GameCommands/Vehicles/VehicleReverse.h"
+#include "GameCommands/Vehicles/VehicleSell.h"
+#include "GameCommands/Vehicles/VehicleSpeedControl.h"
 #include "Graphics/Colour.h"
 #include "Graphics/ImageIds.h"
 #include "Input.h"
